moved WithState and withState to helpers
