+++ src/client/client_api.js
@@ -1,9 +1,10 @@
-import React, { Component, PropTypes } from 'react'
+import { withState, WithState } from './helpers';
 
 export default class ClientApi {
   constructor({ syncedStore, storyStore }) {
     this._syncedStore = syncedStore;
     this._storyStore = storyStore;
+    this.withState = withState;
     this.WithState = WithState;
   }
 
@@ -42,37 +43,5 @@ export default class ClientApi {
       syncedStore.setData({ actions });
     };
   }
+  
 }
-
-
-class WithState extends Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  render() {
-    let { handlers } = this.props;
-    let child = this.props.children;
-
-    let props = Object.assign({}, this.state);
-
-    for (let act in handlers) {
-      let prop = handlers[ act ];
-      props[ act ] = (v) => {
-        if (typeof(child.props[ act ]) === 'function')  {
-          child.props[ act ](v);
-        }
-        this.setState({ [ prop ] : v });
-      }
-    }
-    return React.cloneElement(child, props);
-  }
-
-}
-
-
-WithState.propTypes = {
-  children: PropTypes.object.isRequired
-};
